requirements.txt
django
google-genai
python-dotenv
openai